change align scheduling
