+++ package.json
@@ -75,6 +75,7 @@
     "ngx-infinite-scroll": "10.0.1",
     "ngx-perfect-scrollbar": "^10.1.1",
     "ngx-webstorage": "9.0.0",
+    "observable-webworker": "^4.0.1",
     "perfect-scrollbar": "^1.5.5",
     "primeflex": "^3.1.3",
     "primeicons": "^5.0.0",
