+++ src/main/webapp/app/opencv.worker.ts
@@ -47,10 +47,6 @@ addEventListener('message', e => {
 
       break;
     }
-    case 'imageProcessing':
-      return imageProcessing(e.data);
-    case 'imageAlignement':
-      return imageAlignement(e.data);
     case 'imageCrop':
       return imageCrop(e.data);
     case 'nameprediction':
@@ -63,38 +59,6 @@ addEventListener('message', e => {
       break;
   }
 });
-
-/**
- * With OpenCV we have to work the images as cv.Mat (matrices),
- * so the first thing we have to do is to transform the
- * ImageData to a type that openCV can recognize.
- */
-function imageProcessing(p: { msg: any; payload: any; uid: string }): void {
-  const img = cv.matFromImageData(p.payload);
-  let result = new cv.Mat();
-
-  // What this does is convert the image to a grey scale.
-  cv.cvtColor(img, result, cv.COLOR_BGR2GRAY);
-  postMessage({ msg: p.msg, payload: imageDataFromMat(result), uid: p.uid });
-  img.delete();
-  result.delete();
-}
-
-/**
- * With OpenCV we have to work the images as cv.Mat (matrices),
- * so the first thing we have to do is to transform the
- * ImageData to a type that openCV can recognize.
- */
-function imageAlignement(p: { msg: any; payload: any; uid: string }): void {
-  // What this does is convert the image to a grey scale.
-  if (p.payload.marker) {
-    const resultat = alignImageBasedOnCircle(p.payload);
-    postMessage({ msg: p.msg, payload: resultat, uid: p.uid });
-  } else {
-    const resultat = alignImage(p.payload.imageA, p.payload.imageB, false, 5, 0);
-    postMessage({ msg: p.msg, payload: resultat, uid: p.uid });
-  }
-}
 
 function imageCrop(p: { msg: any; payload: any; uid: string }): void {
   // You can try more different parameters
@@ -188,745 +152,6 @@ function imageDataFromMat(mat: any): any {
   img.delete();
   return clampedArray;
 }
-function alignImageBasedOnCircle(payload: any): any {
-  let srcMat = cv.matFromImageData(payload.imageA);
-  let dst = new cv.Mat(); ///cv.Mat.zeros(srcMat.rows, srcMat.cols, cv.CV_8U);
-  // let color = new cv.Scalar(255, 0, 0);
-  // let displayMat = srcMat.clone();
-  let circlesMat = new cv.Mat();
-  cv.cvtColor(srcMat, srcMat, cv.COLOR_RGBA2GRAY);
-  //  cv.HoughCircles(srcMat, circlesMat, cv.HOUGH_GRADIENT, 1, 45, 75, 40, 0, 0);
-  let minCircle = (srcMat.cols * 6) / 1000;
-  let maxCircle = (srcMat.cols * 20) / 1000;
-
-  cv.HoughCircles(srcMat, circlesMat, cv.HOUGH_GRADIENT, 1, 45, 75, 20, minCircle, maxCircle);
-  let x1, y1, r1;
-  let x2, y2, r2;
-  let x3, y3, r3;
-  let x4, y4, r4;
-  if (circlesMat.cols > 0) {
-    x1 = circlesMat.data32F[0];
-    y1 = circlesMat.data32F[1];
-    r1 = circlesMat.data32F[2];
-    x2 = circlesMat.data32F[0];
-    y2 = circlesMat.data32F[1];
-    r2 = circlesMat.data32F[2];
-    x3 = circlesMat.data32F[0];
-    y3 = circlesMat.data32F[1];
-    r3 = circlesMat.data32F[2];
-    x4 = circlesMat.data32F[0];
-    y4 = circlesMat.data32F[1];
-    r4 = circlesMat.data32F[2];
-  }
-
-  const srcMWidth = srcMat.size().width;
-  const srcMHeight = srcMat.size().height;
-  if (circlesMat.cols > 0) {
-    for (let i = 1; i < circlesMat.cols; i++) {
-      let x = circlesMat.data32F[i * 3];
-      let y = circlesMat.data32F[i * 3 + 1];
-      let radius = circlesMat.data32F[i * 3 + 2];
-      if (x * x + y * y <= x1 * x1 + y1 * y1) {
-        x1 = x;
-        y1 = y;
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        r1 = radius;
-      }
-      if (x * x + y * y >= x4 * x4 + y4 * y4) {
-        x4 = x;
-        y4 = y;
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        r4 = radius;
-      }
-      if ((srcMWidth - x) * (srcMWidth - x) + y * y <= (srcMWidth - x2) * (srcMWidth - x2) + y2 * y2) {
-        x2 = x;
-        y2 = y;
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        r2 = radius;
-      }
-      if (x * x + (srcMHeight - y) * (srcMHeight - y) <= x3 * x3 + (srcMHeight - y3) * (srcMHeight - y3)) {
-        x3 = x;
-        y3 = y;
-        r3 = radius;
-      }
-    }
-  }
-
-  let srcMat2 = cv.matFromImageData(payload.imageB);
-  let srcMat1 = new cv.Mat();
-  let circlesMat1 = new cv.Mat();
-  cv.cvtColor(srcMat2, srcMat1, cv.COLOR_RGBA2GRAY);
-  const srcMWidth1 = srcMat1.size().width;
-  const srcMHeight1 = srcMat1.size().height;
-  //  cv.HoughCircles(srcMat, circlesMat, cv.HOUGH_GRADIENT, 1, 45, 75, 40, 0, 0);
-  cv.HoughCircles(srcMat1, circlesMat1, cv.HOUGH_GRADIENT, 1, 45, 75, 15, r3 - 3, r3 + 3);
-  let goodpointsx = [];
-  let goodpointsy = [];
-  const seuil = ((4 * r3 * r3 - 3.14159 * r3 * r3) * 180) / 100;
-  for (let i = 0; i < circlesMat1.cols; i++) {
-    let x = circlesMat1.data32F[i * 3];
-    let y = circlesMat1.data32F[i * 3 + 1];
-
-    const xx1 = x - r3;
-    const yy1 = y - r3;
-    let width1 = 2 * r3;
-    let height1 = 2 * r3;
-    if (xx1 + width1 > srcMWidth1) {
-      width1 = srcMWidth1 - xx1;
-    }
-    if (yy1 + height1 > srcMHeight1) {
-      height1 = srcMHeight1 - yy1;
-    }
-
-    let rect1 = new cv.Rect(x - r3, y - r3, width1, height1);
-    let dstrect1 = new cv.Mat();
-    dstrect1 = roi(srcMat1, rect1, dstrect1);
-    cv.threshold(dstrect1, dstrect1, 0, 255, cv.THRESH_OTSU + cv.THRESH_BINARY);
-    if (cv.countNonZero(dstrect1) < seuil) {
-      goodpointsx.push(x);
-      goodpointsy.push(y);
-    }
-    dstrect1.delete();
-  }
-
-  let x5, y5;
-  let x6, y6;
-  let x7, y7;
-  let x8, y8;
-  if (goodpointsx.length > 0) {
-    x5 = goodpointsx[0];
-    y5 = goodpointsy[1];
-    x6 = goodpointsx[0];
-    y6 = goodpointsy[1];
-    x7 = goodpointsx[0];
-    y7 = goodpointsy[1];
-    x8 = goodpointsx[0];
-    y8 = goodpointsy[1];
-  }
-
-  if (goodpointsx.length > 1) {
-    for (let i = 0; i < goodpointsx.length; i++) {
-      let x = goodpointsx[i];
-      let y = goodpointsy[i];
-      if (x * x + y * y <= x5 * x5 + y5 * y5) {
-        x5 = x;
-        y5 = y;
-      }
-      if (x * x + y * y >= x8 * x8 + y8 * y8) {
-        x8 = x;
-        y8 = y;
-      }
-      if ((srcMWidth1 - x) * (srcMWidth1 - x) + y * y <= (srcMWidth1 - x6) * (srcMWidth1 - x6) + y6 * y6) {
-        x6 = x;
-        y6 = y;
-      }
-      if (x * x + (srcMHeight1 - y) * (srcMHeight1 - y) <= x7 * x7 + (srcMHeight1 - y7) * (srcMHeight1 - y7)) {
-        x7 = x;
-        y7 = y;
-      }
-    }
-  }
-
-  if (goodpointsx.length >= 4) {
-    let srcTri = cv.matFromArray(4, 1, cv.CV_32FC2, [x1, y1, x2, y2, x3, y3, x4, y4]);
-    let dstTri = cv.matFromArray(4, 1, cv.CV_32FC2, [x5, y5, x6, y6, x7, y7, x8, y8]);
-    //  let srcTri = cv.matFromArray(3, 1, cv.CV_32FC2, [x4, y4, x2, y2, x3, y3]);
-    //  let dstTri = cv.matFromArray(3, 1, cv.CV_32FC2, [x8, y8, x6, y6, x7, y7]);
-    /*  let M = cv.getPerspectiveTransform(dstTri,srcTri);
-//  let M = cv.getAffineTransform(srcTri, dstTri);
-//  console.log(M)
-  let dsize = new cv.Size(srcMat1.rows, srcMat1.cols);
-
-  cv.warpPerspective(srcMat1, dst, M, dsize, cv.BORDER_CONSTANT , cv.BORDER_REPLICATE, new cv.Scalar());*/
-
-    //59    Find homography
-    //60    h = findHomography( points1, points2, RANSAC );
-    //    let h = cv.findHomography(dstTri, srcTri, cv.RANSAC);
-    //    let dsize = new cv.Size(srcMat.cols, srcMat.rows);
-
-    /*   if (h.empty()) {
-      console.log('homography matrix empty!');
-      return;
-    }*/
-    //    let mat1 = cv.matFromArray(4, 1, cv.CV_32FC2, points1);
-    //   let mat2 = cv.matFromArray(4, 1, cv.CV_32FC2,points2);
-    let M = cv.getPerspectiveTransform(dstTri, srcTri);
-    let dsize = new cv.Size(srcMat.cols, srcMat.rows);
-    for (let i = 0; i < srcTri.rows; ++i) {
-      //      let x = srcTri.data32F[i * 2];
-      //      let y = srcTri.data32F[i * 2 + 1];
-      const xx = dstTri.data32F[i * 2];
-      const yy = dstTri.data32F[i * 2 + 1];
-      let radius = 15;
-      let center = new cv.Point(xx, yy);
-      //      cv.circle(srcMat, center, radius, [0, 0, 255, 255], 1);
-      cv.circle(srcMat2, center, radius, [0, 0, 255, 255], 1);
-    }
-
-    cv.warpPerspective(srcMat2, dst, M, dsize, cv.INTER_LINEAR, cv.BORDER_CONSTANT, new cv.Scalar());
-
-    let dst1 = dst.clone();
-
-    /*    for (let i = 0; i < srcTri.rows; ++i) {
-      let x = srcTri.data32F[i * 2];
-      let y = srcTri.data32F[i * 2 + 1];
-      let radius = 15;
-      let center = new cv.Point(x, y);
-//      cv.circle(srcMat2, center, radius, [0, 0, 255, 255], 1);
-    } */
-
-    /*
-let point1 = new cv.Point(payload.x, payload.y);
-let point2 = new cv.Point(payload.x + payload.width, payload.y + payload.height);
-cv.rectangle(dst1, point1, point2,  [0, 0, 255, 255], 2, cv.LINE_AA, 0);
-cv.rectangle(srcMat, point1, point2,  [0, 0, 255, 255], 2, cv.LINE_AA, 0);*/
-
-    let result = {} as any;
-    /*result['keypoints1'] = imageDataFromMat(srcMat);
-    result['keypoints1Width'] = srcMat.size().width;
-    result['keypoints1Height'] = srcMat.size().height;
-    result['keypoints2'] = imageDataFromMat(dst);
-    result['keypoints2Width'] = dst.size().width;
-    result['keypoints2Height'] = dst.size().height;
-    result['imageCompareMatches'] = imageDataFromMat(srcMat2);
-    result['imageCompareMatchesWidth'] = srcMat2.size().width;
-    result['imageCompareMatchesHeight'] = srcMat2.size().height;*/
-    result['imageAligned'] = imageDataFromMat(dst1);
-    result['imageAlignedWidth'] = dst1.size().width;
-    result['imageAlignedHeight'] = dst1.size().height;
-
-    srcMat.delete();
-    dst.delete();
-    circlesMat.delete();
-    srcMat1.delete();
-    srcMat2.delete();
-    circlesMat1.delete();
-    srcTri.delete();
-    dstTri.delete();
-    dst1.delete();
-
-    return result;
-  } else {
-    srcMat.delete();
-    dst.delete();
-    circlesMat.delete();
-    srcMat1.delete();
-    srcMat2.delete();
-    circlesMat1.delete();
-    alignImage(payload.imageB, payload.imageB, false, 5, 0);
-  }
-}
-
-function alignImage(image_A: any, image_B: any, allimage: boolean, numberofpointToMatch: number, numberofgoodpointToMatch: number): any {
-  //im2 is the original reference image we are trying to align to
-  let im2 = cv.matFromImageData(image_A);
-  let im1 = cv.matFromImageData(image_B);
-  let im1Gray = new cv.Mat();
-  let im2Gray = new cv.Mat();
-  cv.cvtColor(im1, im1Gray, cv.COLOR_BGRA2GRAY);
-  cv.cvtColor(im2, im2Gray, cv.COLOR_BGRA2GRAY);
-  const squareSizeorigin = Math.trunc(im2.size().width / 20);
-  let points1: any[] = [];
-  let points2: any[] = [];
-  let result: any = {};
-
-  if (!allimage) {
-    let res = false;
-    let squareSize = squareSizeorigin;
-    while (!res && squareSize < im1Gray.size().width && squareSize < im1Gray.size().height) {
-      let zone1 = new cv.Rect(0, 0, squareSize, squareSize);
-      res = matchSmallImage(im1Gray, im2Gray, points1, points2, zone1, 1, numberofpointToMatch, numberofgoodpointToMatch);
-      if (!res) {
-        squareSize = squareSize + Math.trunc(im2.size().width / 20);
-      }
-    }
-
-    res = false;
-    squareSize = squareSizeorigin;
-    while (!res && squareSize < im1Gray.size().width && squareSize < im1Gray.size().height) {
-      let zone2 = new cv.Rect(im1Gray.size().width - squareSize, 0, im1Gray.size().width, squareSize);
-      res = matchSmallImage(im1Gray, im2Gray, points1, points2, zone2, 2, numberofpointToMatch, numberofgoodpointToMatch);
-      if (!res) {
-        squareSize = squareSize + Math.trunc(im2.size().width / 20);
-      }
-    }
-
-    res = false;
-    squareSize = squareSizeorigin;
-    while (!res && squareSize < im1Gray.size().width && squareSize < im1Gray.size().height) {
-      let zone3 = new cv.Rect(0, im1Gray.size().height - squareSize, squareSize, im1Gray.size().height);
-      res = matchSmallImage(im1Gray, im2Gray, points1, points2, zone3, 3, numberofpointToMatch, numberofgoodpointToMatch);
-
-      if (!res) {
-        squareSize = squareSize + Math.trunc(im2.size().width / 20);
-      }
-    }
-
-    res = false;
-    squareSize = squareSizeorigin;
-    while (!res && squareSize < im1Gray.size().width && squareSize < im1Gray.size().height) {
-      let zone4 = new cv.Rect(
-        im1Gray.size().width - squareSize,
-        im1Gray.size().height - squareSize,
-        im1Gray.size().width,
-        im1Gray.size().height
-      );
-      res = matchSmallImage(im1Gray, im2Gray, points1, points2, zone4, 4, numberofpointToMatch, numberofgoodpointToMatch);
-      if (!res) {
-        squareSize = squareSize + Math.trunc(im2.size().width / 20);
-      }
-    }
-
-    /*
-
-  res = false
-  squareSize =  squareSizeorigin;
-  while (!res && squareSize <   im1Gray.size().width &&  squareSize < im1Gray.size().height){
-    let zone5 = new cv.Rect((im1Gray.size().width -squareSize)/2, (im1Gray.size().height -squareSize)/2, squareSize , squareSize,numberofpointToMatch,numberofgoodpointToMatch);
-    res =  matchSmallImage(im1Gray,im2Gray, points1, points2, zone5,5)
-    if(!res){
-      squareSize = squareSize +Math.trunc(im2.size().width/20);
-    }
-  }*/
-  } else {
-    let zone6 = new cv.Rect(0, 0, im1Gray.size().width, im1Gray.size().height);
-    matchSmallImage(im1Gray, im2Gray, points1, points2, zone6, 1, numberofpointToMatch, numberofgoodpointToMatch);
-  }
-
-  //  console.log('points1:', points1, 'points2:', points2);
-
-  /*  let mat1 = new cv.Mat(points1.length, 1, cv.CV_32FC2);
-  mat1.data32F.set(points1);
-  let mat2 = new cv.Mat(points2.length, 1, cv.CV_32FC2);
-  mat2.data32F.set(points2);
-
-  let h = cv.findHomography(mat1, mat2, cv.RANSAC);
-
-  if (h.empty()) {
-    console.log('homography matrix empty!');
-    return;
-  }
-
-  let image_B_final_result = new cv.Mat();
-  cv.warpPerspective(im1, image_B_final_result, h, im2.size());
-*/
-
-  //  let mat1 = new cv.Mat(points1.length, 1, cv.CV_32FC2);
-  // mat1.data32F.set(points1);
-  // let mat2 = new cv.Mat(points2.length, 1, cv.CV_32FC2);
-  // mat2.data32F.set(points2);
-
-  let mat1 = cv.matFromArray(4, 1, cv.CV_32FC2, points1);
-  let mat2 = cv.matFromArray(4, 1, cv.CV_32FC2, points2);
-  let M = cv.getPerspectiveTransform(mat1, mat2);
-  let image_B_final_result = new cv.Mat();
-
-  cv.warpPerspective(im1, image_B_final_result, M, im2.size(), cv.INTER_LINEAR, cv.BORDER_CONSTANT, new cv.Scalar());
-
-  result['imageAligned'] = imageDataFromMat(image_B_final_result);
-  // result['imageAligned'] = image_B_final_result;
-
-  result['imageAlignedWidth'] = image_B_final_result.size().width;
-  result['imageAlignedHeight'] = image_B_final_result.size().height;
-
-  mat1.delete();
-  mat2.delete();
-  im1.delete();
-  im2.delete();
-
-  im1Gray.delete();
-  im2Gray.delete();
-  return result;
-}
-
-function matchSmallImage(
-  im1Gray: any,
-  im2Gray: any,
-  points1: any[],
-  points2: any[],
-  zone1: any,
-  ii: number,
-  numberofpointToMatch: number,
-  numberofgoodpointToMatch: number
-): boolean {
-  let im1Graydst = new cv.Mat();
-  im1Graydst = roi(im1Gray, zone1, im1Graydst);
-  // cv.imshow("canvas_output2", im1Graydst);
-
-  let im2Graydst = new cv.Mat();
-  im2Graydst = roi(im2Gray, zone1, im2Graydst);
-
-  let keypoints1 = new cv.KeyPointVector();
-  let keypoints2 = new cv.KeyPointVector();
-  let descriptors1 = new cv.Mat();
-  let descriptors2 = new cv.Mat();
-
-  let orb = new cv.AKAZE();
-  let tmp1 = new cv.Mat();
-  let tmp2 = new cv.Mat();
-  orb.detectAndCompute(im1Graydst, tmp1, keypoints1, descriptors1);
-  orb.detectAndCompute(im2Graydst, tmp2, keypoints2, descriptors2);
-
-  let good_matches = new cv.DMatchVector();
-
-  let bf = new cv.BFMatcher();
-  let matches = new cv.DMatchVectorVector();
-  bf.knnMatch(descriptors1, descriptors2, matches, 2);
-
-  //  let counter = 0;
-  for (let i = 0; i < matches.size(); ++i) {
-    let match = matches.get(i);
-    let dMatch1 = match.get(0);
-    let dMatch2 = match.get(1);
-    let knnDistance_option = '0.7';
-    if (dMatch1 !== undefined && dMatch2 !== undefined && dMatch1.distance <= dMatch2.distance * parseFloat(knnDistance_option)) {
-      good_matches.push_back(dMatch1);
-    }
-  }
-
-  // let imMatches = new cv.Mat();
-  // let color = new cv.Scalar(0, 255, 0, 255);
-  // cv.drawMatches(im1, keypoints1, im2, keypoints2, good_matches, imMatches, color);
-  // console.log("good matches "  +  good_matches.size())
-  let points1tmp = [];
-  let points2tmp = [];
-  for (let i = 0; i < good_matches.size(); i++) {
-    points1tmp.push(keypoints1.get(good_matches.get(i).queryIdx).pt.x + zone1.x);
-    points1tmp.push(keypoints1.get(good_matches.get(i).queryIdx).pt.y + zone1.y);
-    points2tmp.push(keypoints2.get(good_matches.get(i).trainIdx).pt.x + zone1.x);
-    points2tmp.push(keypoints2.get(good_matches.get(i).trainIdx).pt.y + zone1.y);
-  }
-  let goodmatchtokeep = 0;
-  let distances = [];
-
-  if (good_matches.size() === 0) {
-    return false;
-  }
-
-  for (let i = 0; i < good_matches.size(); i++) {
-    let distancesquare =
-      (points1tmp[2 * i] - points2tmp[2 * i]) * (points1tmp[2 * i] - points2tmp[2 * i]) +
-      (points1tmp[2 * i + 1] - points2tmp[2 * i + 1]) * (points1tmp[2 * i + 1] - points2tmp[2 * i + 1]);
-    // TODO compute average points
-    if (distancesquare < Math.trunc((3 * im1Graydst.size().width) / 20) * Math.trunc((3 * im1Graydst.size().width) / 20)) {
-      distances.push(distancesquare);
-      goodmatchtokeep = goodmatchtokeep + 1;
-    }
-  }
-
-  const distance = numAverage(distances);
-  const devs = dev(distances);
-
-  if (goodmatchtokeep <= numberofpointToMatch) {
-    orb.delete();
-    keypoints1.delete();
-    keypoints2.delete();
-    descriptors1.delete();
-    descriptors2.delete();
-    tmp1.delete();
-    tmp2.delete();
-    // im1Graydst.delete();
-    im2Graydst.delete();
-    matches.delete();
-    good_matches.delete();
-    bf.delete();
-    return false;
-  }
-  let realgoodmatchtokeep = 0;
-
-  let good_matchesToKeep = [];
-  for (let i = 0; i < good_matches.size(); i++) {
-    let distancesquare =
-      (points1tmp[2 * i] - points2tmp[2 * i]) * (points1tmp[2 * i] - points2tmp[2 * i]) +
-      (points1tmp[2 * i + 1] - points2tmp[2 * i + 1]) * (points1tmp[2 * i + 1] - points2tmp[2 * i + 1]);
-    /*  console.log('distancesquare ' +distancesquare)
-  console.log('distance ' +distance)
-  console.log('devs ' + devs)   */
-    // TODO compute average points
-    if (distancesquare < distance + 1 * devs && distancesquare > distance - 1 * devs) {
-      realgoodmatchtokeep = realgoodmatchtokeep + 1;
-      good_matchesToKeep.push(i);
-      break;
-    }
-  }
-
-  /*let realgoodmatchtokeep1 = new cv.DMatchVector();
-
-//  let counter = 0;
-for (let i = 0; i < good_matches.size(); ++i) {
-  if(good_matchesToKeep.includes(i)){
-    realgoodmatchtokeep1.push_back(good_matches.get(i));
-  }
-}
-
-if(realgoodmatchtokeep >0) {
-  let imMatches = new cv.Mat();
-  let color = new cv.Scalar(0, 255, 0, 255);
-  cv.drawMatches(im1Graydst, keypoints1, im2Graydst, keypoints2, realgoodmatchtokeep1, imMatches, color);
-  cv.imshow("canvas_output" + (1+i), imMatches);
-
-}*/
-
-  // console.log(realgoodmatchtokeep)
-  if (realgoodmatchtokeep <= numberofgoodpointToMatch) {
-    orb.delete();
-    keypoints1.delete();
-    keypoints2.delete();
-    descriptors1.delete();
-    descriptors2.delete();
-    tmp1.delete();
-    tmp2.delete();
-    // im1Graydst.delete();
-    im2Graydst.delete();
-    matches.delete();
-    good_matches.delete();
-    bf.delete();
-    return false;
-  } else {
-    good_matchesToKeep.forEach(i => {
-      points1.push(keypoints1.get(good_matches.get(+i).queryIdx).pt.x + zone1.x);
-      points1.push(keypoints1.get(good_matches.get(+i).queryIdx).pt.y + zone1.y);
-      points2.push(keypoints2.get(good_matches.get(+i).trainIdx).pt.x + zone1.x);
-      points2.push(keypoints2.get(good_matches.get(+i).trainIdx).pt.y + zone1.y);
-      /* let radius = 15;
-      let center = new cv.Point(keypoints1.get(good_matches.get(i).queryIdx).pt.x+ zone1.x, keypoints1.get(good_matches.get(i).queryIdx).pt.y+ zone1.y);
-      let center1 = new cv.Point(keypoints2.get(good_matches.get(i).trainIdx).pt.x+ zone1.x, keypoints2.get(good_matches.get(i).trainIdx).pt.y+ zone1.y);
-      cv.circle(im1Gray, center, radius, [0, 0, 255, 255], 1);
-      cv.circle(im2Gray, center1, radius, [0, 0, 255, 255], 1);
-      cv.imshow("canvas_output6", im2Gray);
-      cv.imshow("canvas_output7" , im1Gray); */
-    });
-    orb.delete();
-    keypoints1.delete();
-    keypoints2.delete();
-    descriptors1.delete();
-    descriptors2.delete();
-    tmp1.delete();
-    tmp2.delete();
-    // im1Graydst.delete();
-    im2Graydst.delete();
-    matches.delete();
-    good_matches.delete();
-    bf.delete();
-
-    console.log('good match for zone ' + ii + ' = ' + realgoodmatchtokeep);
-
-    return true;
-  }
-}
-
-function numAverage(arr: number[]): number {
-  return (
-    arr.reduce(
-      (acc: number, curr: number) => acc + curr,
-
-      0
-    ) / arr.length
-  );
-}
-
-// Javascript program to calculate the standered deviation of an array
-function dev(arr: number[]): number {
-  // Creating the mean with Array.reduce
-  let mean = arr.reduce((acc, curr) => acc + curr, 0) / arr.length;
-
-  // Assigning (value - mean) ^ 2 to every array item
-  arr = arr.map(k => (k - mean) ** 2);
-
-  // Calculating the sum of updated array
-  let sum = arr.reduce((acc, curr) => acc + curr, 0);
-
-  // Calculating the variance
-  // let variance = sum / arr.length
-
-  // Returning the Standered deviation
-  return Math.sqrt(sum / arr.length);
-}
-
-/*
-function alignImage(image_A: any, image_B: any): any {
-  //im2 is the original reference image we are trying to align to
-  let im2 = cv.matFromImageData(image_A);
-  let im1 = cv.matFromImageData(image_B);
-  let im1Gray = new cv.Mat();
-  let im2Gray = new cv.Mat();
-  cv.cvtColor(im1, im1Gray, cv.COLOR_BGRA2GRAY);
-  cv.cvtColor(im2, im2Gray, cv.COLOR_BGRA2GRAY);
-
-  //22    Variables to store keypoints and descriptors
-  //23    std::vector<KeyPoint> keypoints1, keypoints2;
-  //24    Mat descriptors1, descriptors2;
-  let keypoints1 = new cv.KeyPointVector();
-  let keypoints2 = new cv.KeyPointVector();
-  let descriptors1 = new cv.Mat();
-  let descriptors2 = new cv.Mat();
-
-  //26    Detect ORB features and compute descriptors.
-  //27    Ptr<Feature2D> orb = ORB::create(MAX_FEATURES);
-  //28    orb->detectAndCompute(im1Gray, Mat(), keypoints1, descriptors1);
-  //29    orb->detectAndCompute(im2Gray, Mat(), keypoints2, descriptors2);
-
-  let orb = new cv.AKAZE();
-  let tmp1 = new cv.Mat();
-  let tmp2 = new cv.Mat();
-  orb.detectAndCompute(im1Gray, tmp1, keypoints1, descriptors1);
-  orb.detectAndCompute(im2Gray, tmp2, keypoints2, descriptors2);
-
-  /*console.log('Total of ', keypoints1.size(), ' keypoints1 (img to align) and ', keypoints2.size(), ' keypoints2 (reference)');
-  console.log('here are the first 5 keypoints for keypoints1:');
-  for (let i = 0; i < keypoints1.size(); i++) {
-    console.log('keypoints1: [', i, ']', keypoints1.get(i).pt.x, keypoints1.get(i).pt.y);
-    if (i === 5) {
-      break;
-    }
-  }*/
-
-//31    Match features.
-//32    std::vector<DMatch> matches;
-//33    Ptr<DescriptorMatcher> matcher = DescriptorMatcher::create("BruteForce-Hamming");
-//34    matcher->match(descriptors1, descriptors2, matches, Mat());
-
-/*  let good_matches = new cv.DMatchVector();
-
-  //36    Sort matches by score
-  //37    std::sort(matches.begin(), matches.end());
-
-  //39    Remove not so good matches
-  //40    const int numGoodMatches = matches.size() * GOOD_MATCH_PERCENT;
-  //41    matches.erase(matches.begin()+numGoodMatches, matches.end());
-
-  let bf = new cv.BFMatcher();
-  let matches = new cv.DMatchVectorVector();
-  bf.knnMatch(descriptors1, descriptors2, matches, 2);
-
-  //  let counter = 0;
-  for (let i = 0; i < matches.size(); ++i) {
-    let match = matches.get(i);
-    let dMatch1 = match.get(0);
-
-    let dMatch2 = match.get(1);
-    // console.log("[", i, "] ", "dMatch1: ", dMatch1, "dMatch2: ", dMatch2);
-    let knnDistance_option = '0.7';
-    if (dMatch1.distance <= dMatch2.distance * parseFloat(knnDistance_option)) {
-      //console.log("***Good Match***", "dMatch1.distance: ", dMatch1.distance, "was less than or = to: ", "dMatch2.distance * parseFloat(knnDistance_option)", dMatch2.distance * parseFloat(knnDistance_option), "dMatch2.distance: ", dMatch2.distance, "knnDistance", knnDistance_option);
-      good_matches.push_back(dMatch1);
-      //     counter++;
-    }
-  }
-
-  /*  console.log('keeping ', counter, ' points in good_matches vector out of ', matches.size(), ' contained in this match vector:', matches);
-  console.log('here are first 5 matches');
-  for (let t = 0; t < matches.size(); ++t) {
-    console.log('[' + t + ']', 'matches: ', matches.get(t));
-    if (t === 5) {
-      break;
-    }
-  }
-
-  console.log('here are first 5 good_matches');
-  for (let r = 0; r < good_matches.size(); ++r) {
-    console.log('[' + r + ']', 'good_matches: ', good_matches.get(r));
-    if (r === 5) {
-      break;
-    }
-  }
-*/
-//44    Draw top matches
-//45    Mat imMatches;
-//46    drawMatches(im1, keypoints1, im2, keypoints2, matches, imMatches);
-//47    imwrite("matches.jpg", imMatches);
-/* let imMatches = new cv.Mat();
-  let color = new cv.Scalar(0, 255, 0, 255);
-  cv.drawMatches(im1, keypoints1, im2, keypoints2, good_matches, imMatches, color);
-  // TODO pass the canvas
-
-  let result = {} as any;
-
-  result['imageCompareMatches'] = imageDataFromMat(imMatches);
-  result['imageCompareMatchesWidth'] = imMatches.size().width;
-  result['imageCompareMatchesHeight'] = imMatches.size().height;
-
-  let keypoints1_img = new cv.Mat();
-  let keypoints2_img = new cv.Mat();
-  let keypointcolor = new cv.Scalar(0, 255, 0, 255);
-  cv.drawKeypoints(im1Gray, keypoints1, keypoints1_img, keypointcolor);
-  cv.drawKeypoints(im2Gray, keypoints2, keypoints2_img, keypointcolor);
-
-  result['keypoints1'] = imageDataFromMat(keypoints1_img);
-  result['keypoints1Width'] = keypoints1_img.size().width;
-  result['keypoints1Height'] = keypoints1_img.size().height;
-  result['keypoints2'] = imageDataFromMat(keypoints2_img);
-  result['keypoints2Width'] = keypoints2_img.size().width;
-  result['keypoints2Height'] = keypoints2_img.size().height;
-
-  //50    Extract location of good matches
-  //51    std::vector<Point2f> points1, points2;
-  //53    for( size_t i = 0; i < matches.size(); i++ )
-  //54    {
-  //55        points1.push_back( keypoints1[ matches[i].queryIdx ].pt );
-  //56        points2.push_back( keypoints2[ matches[i].trainIdx ].pt );
-  //57    }
-  let points1 = [];
-  let points2 = [];
-  for (let i = 0; i < good_matches.size(); i++) {
-    points1.push(keypoints1.get(good_matches.get(i).queryIdx).pt.x);
-    points1.push(keypoints1.get(good_matches.get(i).queryIdx).pt.y);
-    points2.push(keypoints2.get(good_matches.get(i).trainIdx).pt.x);
-    points2.push(keypoints2.get(good_matches.get(i).trainIdx).pt.y);
-  }
-
-  //  console.log('points1:', points1, 'points2:', points2);
-
-  let mat1 = new cv.Mat(points1.length, 1, cv.CV_32FC2);
-  mat1.data32F.set(points1);
-  let mat2 = new cv.Mat(points2.length, 1, cv.CV_32FC2);
-  mat2.data32F.set(points2);
-
-  //59    Find homography
-  //60    h = findHomography( points1, points2, RANSAC );
-  let h = cv.findHomography(mat1, mat2, cv.RANSAC);
-
-  if (h.empty()) {
-    console.log('homography matrix empty!');
-    return;
-  } else {
-    console.log('h:', h);
-    //    console.log('[', h.data64F[0], ',', h.data64F[1], ',', h.data64F[2]);
-    //    console.log('', h.data64F[3], ',', h.data64F[4], ',', h.data64F[5]);
-    //    console.log('', h.data64F[6], ',', h.data64F[7], ',', h.data64F[8], ']');
-  }
-
-  //62  Use homography to warp image
-  //63  warpPerspective(im1, im1Reg, h, im2.size());
-
-  let image_B_final_result = new cv.Mat();
-  cv.warpPerspective(im1, image_B_final_result, h, im2.size());
-
-  result['imageAligned'] = imageDataFromMat(image_B_final_result);
-  result['imageAlignedWidth'] = image_B_final_result.size().width;
-  result['imageAlignedHeight'] = image_B_final_result.size().height;
-
-  image_B_final_result.delete();
-  mat1.delete();
-  mat2.delete();
-  keypoints1_img.delete();
-  keypoints2_img.delete();
-  imMatches.delete();
-  matches.delete();
-  bf.delete();
-  good_matches.delete();
-  orb.delete();
-  im1Gray.delete();
-  keypoints1.delete();
-  im2Gray.delete();
-  keypoints2.delete();
-  descriptors1.delete();
-  descriptors2.delete();
-  tmp1.delete();
-  tmp2.delete();
-  //h.delete();
-  return result;
-}*/
 
 function fprediction(src: any, cand: string[], m: MLModel, lookingForMissingLetter: boolean, onlyletter: boolean): any {
   const res = extractImage(src, false, lookingForMissingLetter);
